+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.42.5] - 2025-07-01
 ### Improved
 - **qol**: Add exports for `JWTError`, `JWTMalformedError`, `JWTTypeError`, `JWTTimeError`, `JWTClaimError`, `JWTAlgorithmError`, `JWTSignatureError`
 - **qol**: Make `algorithm` option when verifying JWT optional and default to `HS256` (the same default as when signing)